# laravel/source/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/image": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "Get All Image",
                "operationId": "a9ebfd028263c21d5828dc0b18579234",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "Add Image",
                "operationId": "eee9961c7d41fa294a590e08fd5e9296",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "type": "string"
                                    },
                                    "varianId": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/image/{id}": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "Get Image",
                "operationId": "c343f5b4d8a1a540c30568616b6ffd4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Image"
                ],
                "summary": "Update Image",
                "operationId": "8762af91bafd1a93ef3d7c0b1cd8d75d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "varianId": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Image"
                ],
                "summary": "Delete Image",
                "operationId": "405d5dc41ba6f72ffc5ed8fcf33bf9d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/productcollection": {
            "get": {
                "tags": [
                    "ProductCollection"
                ],
                "summary": "Get All ProductCollection",
                "operationId": "12513af85c21b4dd765f2885c6d7c2b9",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "ProductCollection"
                ],
                "summary": "Add ProductCollection",
                "operationId": "3fab329f5595daae6c859c8f7dd18cb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "productId": {
                                        "type": "integer"
                                    },
                                    "collectionId": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/productcollection/{id}": {
            "get": {
                "tags": [
                    "ProductCollection"
                ],
                "summary": "Get ProductCollection",
                "operationId": "7810c761033149a9193516e0ae40f1b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductCollection"
                ],
                "summary": "Update ProductCollection",
                "operationId": "d2ef057dc5d01bbfc38b3bfc648ea454",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "productId": {
                                        "type": "integer"
                                    },
                                    "collectionId": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductCollection"
                ],
                "summary": "Delete ProductCollection",
                "operationId": "33cc7d8c0e45e0c7b21ce2f27e9ebbb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}